Changed timer to be more accurate

+++ Day01/Day1.java
@@ -41,7 +41,8 @@ public class Day1 {
             System.out.println(e.toString());
         }
         long endTime = Benchmark.currentTime();
-        long elapsed = Benchmark.elapsedTime(startTime, endTime);
+        
+        double elapsed = Benchmark.elapsedTime(startTime, endTime);
      
         System.out.println("Part 1: " + answerPart1);
         System.out.println("Part 2: " + answerPart2);
