+++ Day09/Day9.java
@@ -38,7 +38,7 @@ public class Day9 {
             System.out.println(e.toString());
         }
         long endTime = Benchmark.currentTime();
-        long elapsed = Benchmark.elapsedTime(startTime, endTime);
+        double elapsed = Benchmark.elapsedTime(startTime, endTime);
      
         System.out.println("Part 1: " + answerPart1);
         System.out.println("Part 2: " + answerPart2);
