+++ Day10/Day10.java
@@ -85,7 +85,7 @@ public class Day10 {
         floodAnswerPart2 = interiorTiles / 9;
         
         long endTime = Benchmark.currentTime();
-        long elapsed = Benchmark.elapsedTime(startTime, endTime);
+        double elapsed = Benchmark.elapsedTime(startTime, endTime);
      
         System.out.println("Part 1: " + answerPart1);
         System.out.println("Part 2 parity method: " + parityAnswerPart2);
