+++ Day13/Day13.java
@@ -65,7 +65,7 @@ public class Day13 {
         }
         
         long endTime = Benchmark.currentTime();
-        long elapsed = Benchmark.elapsedTime(startTime, endTime);
+        double elapsed = Benchmark.elapsedTime(startTime, endTime);
      
         System.out.println("Part 1: " + answerPart1);
         System.out.println("Part 2: " + answerPart2);
