+++ Day20/Day20.java
@@ -30,7 +30,7 @@ public class Day20 {
         }  
         
         long endTime = Benchmark.currentTime();
-        long elapsed = Benchmark.elapsedTime(startTime, endTime);
+        double elapsed = Benchmark.elapsedTime(startTime, endTime);
      
         System.out.println("Part 1: " + answerPart1);
         System.out.println("Part 2: " + answerPart2);
